requirements.txt
numpy==1.26.4
Pillow>=10.0.0
PyYAML==6.0.1
matplotlib>=3.7.0
scikit-learn>=1.3.0
pandas>=2.0.0
pytest>=7.4.0
torch>=2.1.0
torchvision>=0.16.0
facenet-pytorch>=2.5.3
tqdm>=4.65.0
requests>=2.31.0
